Some initial refactoring
